+++ app/src/main/java/com/whzl/mengbi/model/LiveModel.java
@@ -11,9 +11,9 @@ public interface LiveModel {
     void doRoomInfo(int programId, OnLiveFinishedListener listener);
     void doAudienceAccount(int programId, OnLiveFinishedListener listener);
 
-    void doFollowHost(int userId, int programId, OnLiveFinishedListener listener);
+    void doFollowHost(long userId, int programId, OnLiveFinishedListener listener);
 
-    void doRoomUserInfo(int userId, int programId, OnLiveFinishedListener listener);
+    void doRoomUserInfo(long userId, int programId, OnLiveFinishedListener listener);
 
-    void doSendGift(int userId, int count, int goodsId, int programId, int targetId, OnLiveFinishedListener listener);
+    void doSendGift(long userId, int count, int goodsId, int programId, int targetId, OnLiveFinishedListener listener);
 }
